pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["nlpwantyou <king.god.ant@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.73.0"
sqlmodel = "^0.0.6"
selenium = "^4.1.0"
bs4 = "^0.0.1"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
